2nd Commit - Well Technically not
Had some trouble with git because I forgot how to use it and messed up my commiting but I applied some styles to try to get closer to the lay out of the excercise, having some trouble with the divs in the middle. Even after wrapping a containter Ican't get them to align on top of each other. I'm surely over looking something simple.