Final Commit - Ready to move on
Got a bit lazy and tired of this excercise as I really wanted to move to js fundamentals. used Chatgpt to help me a bit as I was having some trouble with some properties. Particularly I was using divs for everything and I was having trouble tsacking some of them.